move zbase32 to its own module
